chore: fix test on 32bit and non-cgo platforms
This is a pretty common helper library, so we'd like it to work everywhere.
